Necessary fixes after plugins update (#127)
JobDSL plugin removed support for ScriptLocation class https://github.com/jenkinsci/job-dsl-plugin/wiki/Migration#migrating-to-149, and this update broke initialjobs.groovy where that class was used. I rewrote initialjobs.groovy to be a pipeline job instead of freestyle project and updated it not to use ScriptLocation

The Workflow Plugin has been renamed to Pipeline Plugin. New pipelineJob and multibranchPipelineJob methods have been added as replacements for workflowJob and multibranchWorkflowJob which are deprecated and removed now. We had to update JobBuilder to reflect those changes
https://github.com/jenkinsci/job-dsl-plugin/wiki/Migration#migrating-to-147